+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for quoting keys in `YAML` paths to allow for keys with dots (.)
+- Expansion of system properties in YAML
 
 ## [1.4.7](https://github.com/kb-dk/kb-util/tree/kb-util-1.4.7)
 ### Changed